组件抽象

+++ pacifica/src/model/replica_group.rs
@@ -4,8 +4,8 @@ use crate::model::replica_id::ReplicaId;
 pub struct ReplicaGroup {
     pub primary: ReplicaId,
     pub secondaries: Vec<ReplicaId>,
-    pub version: u64,
-    pub term: u64
+    pub version: usize,
+    pub term: usize
 }
 
 impl Display for ReplicaGroup {
@@ -17,3 +17,4 @@ impl Display for ReplicaGroup {
 impl ReplicaGroup {
     
 }
+
